backend/requirements.txt
# Dependências principais
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# IA e Processamento de Imagem
opencv-python==4.8.1.78
mediapipe==0.10.8
Pillow==10.1.0
numpy==1.24.3

# Utilitários
python-decouple==3.8
